+++ CHANGES.rst
@@ -4,6 +4,20 @@
 Changes in current version of Apache Libcloud
 ---------------------------------------------
 
+Compute
+~~~~~~~
+
+- [EC2] add g3 instance types
+  [GITHUB-1101]
+  (@zulupro)
+
+- [EC2] add 'end' to ec2 reserved_node
+  [GITHUB-1099]
+  (@xofer)
+
+Changes in Apache Libcloud 2.2.0
+--------------------------------
+
 Common
 ~~~~~~